requirements.txt
pandas
numpy 
scikit-learn
statistics
imbalanced-learn
pickle5
pyyaml
regex
-e .